Added two columns to section Ice Modifiers.
I thought it would be nice to quickly check what works for plugins and
snippets. It also eliminates any ambiguity of when an Ice modifier can
be used.

Only problem is that I'm not certain for some modifiers so psprint or
someone else may have to fix or correct wrong info.